+++ app-backend/datamodel/src/main/scala/Metric.scala
@@ -1,34 +1,31 @@
 package com.rasterfoundry.datamodel
 
 import java.time.{Instant, LocalDate, LocalDateTime, ZoneOffset}
-import java.util.UUID
 
 import io.circe._
 import io.circe.generic.semiauto._
 
+import java.time.{Instant, LocalDateTime, LocalDate, ZoneOffset}
+
 case class Metric(
-    id: UUID,
     period: (LocalDate, LocalDate),
     metricEvent: MetricEvent,
-    value: Int,
-    requester: String
+    requester: String,
+    value: Int
 )
 
 object Metric {
 
   implicit val encMetric: Encoder[Metric] = deriveEncoder[Metric]
 
-  def apply(id: UUID,
+  def apply(occurredAt: Instant,
             metricEvent: MetricEvent,
-            occurredAt: Instant,
-            value: Int,
             requester: String): Metric =
     Metric(
-      id,
       rangeForInstant(occurredAt),
       metricEvent,
-      value,
-      requester
+      requester,
+      1
     )
 
   def rangeForInstant(instant: Instant): (LocalDate, LocalDate) = {